% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_osmium_is_installed.R
\name{assert_osmium_is_installed}
\alias{assert_osmium_is_installed}
\title{Assert Osmium is installed}
\usage{
assert_osmium_is_installed()
}
\value{
Throws an error if Osmium is not installed, invisibly returns \code{TRUE}
otherwise.
}
\description{
Asserts Osmium is installed, throwing an error otherwise.
}
\keyword{internal}
