% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_info.R
\name{osm_file_info_variables}
\alias{osm_file_info_variables}
\title{List available variables for osm_file_info}
\usage{
osm_file_info_variables(echo_cmd = FALSE, spinner = TRUE, echo = FALSE)
}
\arguments{
\item{echo_cmd}{A logical. Whether to print the Osmium command generated by the function call to the screen.
Defaults to \code{FALSE}.}

\item{spinner}{A logical. Whether to show a reassuring spinner while the Osmium call is being executed.
Defaults to \code{TRUE}.}

\item{echo}{A logical. Whether to print the standard output and error generated by the Osmium call to the screen.
Defaults to \code{TRUE}.}
}
\value{
Character vector of variable names.
}
\description{
Fetches the set of valid variable names supported by the underlying Osmium CLI.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
vars <- osm_file_info_variables()
head(vars)
\dontshow{\}) # examplesIf}
}
