% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bbox.R
\name{osm_set_bbox}
\alias{osm_set_bbox}
\title{Apply or Set Bounding‐Box Header on an OSM File}
\usage{
osm_set_bbox(
  input_path,
  output_path,
  bbox = NULL,
  calculate = TRUE,
  overwrite = FALSE,
  echo_cmd = FALSE,
  echo = TRUE,
  spinner = TRUE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{input_path}{String. Path to the existing OSM file.}

\item{output_path}{String. Path where the new file will be written.}

\item{bbox}{An \code{sf::st_bbox} to use.  If \code{NULL},
the bounding box is computed via \code{osm_get_bbox()}.}

\item{calculate}{Logical(1). If \code{TRUE}, allow computing from data
when no header is present; otherwise error if missing.
Defaults to \code{TRUE}.}

\item{overwrite}{Logical(1). If \code{TRUE}, allow overwriting
\code{output_path}.  Defaults to \code{FALSE}.}

\item{echo_cmd}{A logical. Whether to print the Osmium command generated by the function call to the screen.
Defaults to \code{FALSE}.}

\item{echo}{A logical. Whether to print the standard output and error generated by the Osmium call to the screen.
Defaults to \code{TRUE}.}

\item{spinner}{A logical. Whether to show a reassuring spinner while the Osmium call is being executed.
Defaults to \code{TRUE}.}

\item{verbose}{A logical. Whether to display detailed information on the running command. Defaults to \code{FALSE}.}

\item{progress}{A logical. Whether to display a progress bar while running
the command. Defaults to \code{FALSE}.}
}
\value{
Invisibly returns the normalized path to \code{output_path}.
}
\description{
Writes a new PBF file (or overwrites it) with its header bounding box set
to either a user‐supplied \code{sf::st_bbox} or—by default—the computed data
bounding box.  Computing may require reading the entire file and can be slow!
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pbf <- system.file("extdata/cur.osm.pbf", package = "rosmium")

# Compute true bbox and write to a new file
out <- tempfile(fileext = ".osm.pbf")
osm_set_bbox(pbf, out, calculate = TRUE, overwrite = TRUE)

# Supply a custom bbox
library(sf)
custom <- st_bbox(
  c(xmin = -49.3, ymin = -25.45, xmax = -49.24, ymax = -25.41),
  crs = st_crs(4326)
)
out2 <- tempfile(fileext = ".osm.pbf")
osm_set_bbox(pbf, out2, bbox = custom, overwrite = TRUE)
\dontshow{\}) # examplesIf}
}
