% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_refs.R
\name{print.osm_check_refs_log}
\alias{print.osm_check_refs_log}
\title{Print method for \code{osm_check_refs_log} objects}
\usage{
\method{print}{osm_check_refs_log}(x, ...)
}
\arguments{
\item{x}{An \code{osm_check_refs_log} object}

\item{...}{Additional args (ignored)}
}
\description{
Displays parsed results in a format resembling \verb{osmium check-refs} native output.
}
