\name{Rosmium-package}
\alias{Rosmium-package}
\alias{Rosmium}
\docType{package}
\title{
\packageTitle{Rosmium}
}
\description{
\packageDescription{Rosmium}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rosmium}
\packageIndices{Rosmium}

}
\author{
\packageAuthor{Rosmium}

Maintainer: \packageMaintainer{Rosmium}

}

\references{
See \url{http://osmcode.org/libosmium} for further information about Libosmium.
}

\keyword{ package }
\seealso{
}
\examples{

}
