% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosmium.R
\docType{package}
\name{rosmium}
\alias{rosmium}
\alias{_PACKAGE}
\alias{rosmium-package}
\title{rosmium: Bindings for Osmium Tool}
\description{
Allows one to use Osmium Tool from R. Osmium is a multipurpose command line
tool that enables one to manipulate and analyze OpenStreetMap (OSM) files
through several different commands. Currently, this package does not aim to
offer functions that cover the entirety of Osmium’s API, instead making
available functions that wrap only a very limited set of Osmium’s features.
}
\section{Usage}{

Please check the vignette for more on the package usage:
\itemize{
\item Introductino to rosmium: R bindings for Osmium Tool. Run
\code{vignette("rosmium")} or check it on the
\href{https://ipeagit.github.io/rosmium/articles/rosmium.html}{website}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeaGIT/rosmium}
  \item Report bugs at \url{https://github.com/ipeaGIT/rosmium/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

}
\keyword{internal}
