% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_filter.R
\name{tags_filter}
\alias{tags_filter}
\title{Filter objects matching specified keys/tags}
\usage{
tags_filter(
  input_path,
  filters,
  output_path,
  overwrite = FALSE,
  echo_cmd = FALSE,
  echo = TRUE,
  spinner = TRUE
)
}
\description{
Filter objects matching specified keys/tags
}
\examples{
pbf_path <- system.file("extdata/cur.osm.pbf", package = "rosmium")

}
\keyword{internal}
