% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_osmium_is_installed.R
\name{assert_osmium_is_installed}
\alias{assert_osmium_is_installed}
\title{Assert Osmium is installed}
\usage{
assert_osmium_is_installed()
}
\value{
Throws an error if Osmium is not installed, invisibly returns \code{TRUE}
otherwise.
}
\description{
Asserts Osmium is installed, throwing an error otherwise.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
assert_osmium_is_installed()
\dontshow{\}) # examplesIf}
}
