% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{osm_merge}
\alias{osm_merge}
\title{Merge several sorted OSM files}
\usage{
osm_merge(
  input_paths,
  output_path,
  check_integrity = "no",
  input_format = NULL,
  output_format = NULL,
  fsync = FALSE,
  generator = NULL,
  overwrite = FALSE,
  output_header = NULL,
  echo_cmd = FALSE,
  echo = TRUE,
  spinner = TRUE,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{input_paths}{A character vector of paths to one or more existing, sorted OSM files.}

\item{output_path}{A string. Path to the file where the merged result will be written.}

\item{check_integrity}{A string. One of "no", "simple", or "full". "no" (default) skips integrity checks;
"simple" runs \code{\link[=osm_check_refs]{osm_check_refs()}} without relation checks;
"full" runs \code{\link[=osm_check_refs]{osm_check_refs()}} with relation checks enabled.}

\item{input_format}{A string or \code{NULL}. Force input format (e.g. "osm", "pbf"). If \code{NULL}, autodetect.}

\item{output_format}{A string or \code{NULL}. Force output format (e.g. "osm", "pbf"). If \code{NULL}, autodetect.}

\item{fsync}{A logical. If \code{TRUE}, call \code{fsync} after writing to force flushing buffers. Defaults to \code{FALSE}.}

\item{generator}{A string or \code{NULL}. Name/version for the \code{generator} header. Defaults to Osmium’s own.}

\item{overwrite}{A logical. If \code{TRUE}, allow overwriting an existing \code{output_path}. Defaults to \code{FALSE}.}

\item{output_header}{A named character vector of additional header options. Each name is the header key,
each value is the header value. A zero‐length value (\code{""}) writes \verb{OPTION!}. Defaults to \code{NULL}.}

\item{echo_cmd}{A logical. Whether to print the generated Osmium command. Defaults to \code{FALSE}.}

\item{echo}{A logical. Whether to print Osmium’s stdout/stderr. Defaults to \code{TRUE}.}

\item{spinner}{A logical. Whether to show a spinner during execution. Defaults to \code{TRUE}.}

\item{verbose}{A logical. Whether to pass \code{--verbose} to Osmium for detailed logging. Defaults to \code{FALSE}.}

\item{progress}{A logical. Whether to pass \code{--progress} / \code{--no-progress}. Defaults to \code{FALSE}.}
}
\value{
Invisibly returns the normalized path to \code{output_path}.
}
\description{
This is a wrapper for \href{https://docs.osmcode.org/osmium/latest/osmium-merge.html}{osmium merge}.
Merges multiple sorted OSM files into a single output, removing duplicate objects.
Optionally checks referential integrity of each input before merging.
}
\keyword{internal}
