% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_content.R
\name{show_content}
\alias{show_content}
\title{Show the contents of an OSM file}
\usage{
show_content(
  input_path,
  output_format = c("debug", "opl", "xml"),
  object_type = c("all", "node", "way", "relation", "changeset"),
  echo_cmd = FALSE,
  spinner = TRUE,
  preview = TRUE
)
}
\description{
Show the contents of an OSM file
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
pbf_path <- system.file("extdata/cur.osm.pbf", package = "rosmium")

small_pbf <- tags_filter(pbf_path, "n/note", tempfile(fileext = "osm.pbf"))

rosmium:::show_content(small_pbf)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
